Structuring project adding new folders and PlayerEventHandler.java
