+++ src/main/java/com/blocketing/HttpMinecraftServer.java
@@ -1,10 +1,14 @@
 package com.blocketing;
 
+import com.blocketing.config.ConfigLoader;
+import com.blocketing.discord.DiscordMessageHandler;
 import com.sun.net.httpserver.HttpServer;
 import net.minecraft.server.MinecraftServer;
 import java.net.InetSocketAddress;
 
-// This class would be responsible for starting the HTTP server to listen for messages from Discord.
+/**
+ * This class would be responsible for starting the HTTP server to listen for messages from Discord.
+ */
 public class HttpMinecraftServer {
 
     /**
@@ -22,8 +26,8 @@ public class HttpMinecraftServer {
 
             // Starts the HTTP server on the configured port
             HttpServer server = HttpServer.create(new InetSocketAddress(port), 0);
-            server.createContext("/discord-to-minecraft", exchange -> ChatHandlerDiscord.handleDiscordMessage(exchange, minecraftServer));
-            server.createContext("/execute-command", exchange -> ChatHandlerDiscord.handleCommand(exchange, minecraftServer));
+            server.createContext("/discord-to-minecraft", exchange -> DiscordMessageHandler.handleDiscordMessage(exchange, minecraftServer));
+            server.createContext("/execute-command", exchange -> DiscordMessageHandler.handleCommand(exchange, minecraftServer));
             server.setExecutor(null);
             server.start();
             System.out.println("DiscordHandler listening on port " + port);
